pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "reviewercalls"
version = "2026.2.2"
description = "Automated discovery of reviewer self-nomination calls for CS conferences"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Zhongkui Ma", email = "zhongkuima0419@gmail.com"}]
dependencies = [
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "aiohttp>=3.13.3",
    "validators>=0.35.0",
    "ddgs>=9.10.0",
    "ollama>=0.4.0",
    "json-repair>=0.30.0",
]

[project.optional-dependencies]
dev = ["pytest>=9.0.2", "pytest-asyncio>=1.3.0", "ruff>=0.15.0", "pre-commit>=4.5.1"]

[tool.setuptools]
packages = ["discover", "validate", "build", "tests"]

[tool.pytest.ini_options]
testpaths = ["tests"]
